chore(js): updated the code with suggestions from pr #6362
