+++ packages/js/src/ui/components/elements/Bell/Bell.tsx
@@ -1,6 +1,7 @@
-import { Component, Show, onCleanup, createEffect } from 'solid-js';
+import { Component, Show } from 'solid-js';
 import { useTotalUnreadCount } from '../../../context';
 import { BellRenderer } from '../../../types';
+import { ExternalElementRenderer } from '../../ExternalElementRenderer';
 import { BellContainer } from './DefaultBellContainer';
 
 type BellProps = {
@@ -8,24 +9,11 @@ type BellProps = {
 };
 /* This is also going to be exported as a separate component. Keep it pure. */
 export const Bell: Component<BellProps> = (props) => {
-  let ref: HTMLDivElement;
   const { totalUnreadCount } = useTotalUnreadCount();
-
-  createEffect(() => {
-    const unmount = props.renderBell?.(ref, totalUnreadCount());
-
-    onCleanup(() => {
-      unmount?.();
-    });
-  });
 
   return (
     <Show when={props.renderBell} fallback={<BellContainer unreadCount={totalUnreadCount()} />}>
-      <div
-        ref={(el) => {
-          ref = el;
-        }}
-      />
+      <ExternalElementRenderer render={(el) => props.renderBell!(el, totalUnreadCount())} />
     </Show>
   );
 };
